Revert "Initial implementation for embedded datastore (#500)"
This reverts commit 4036489bfcf6557547bde50d973e95273428f90a.
