--- docs/src/snap/howto/datastore/index.md
@@ -1,23 +1,0 @@
-# Datastore
-
-```{toctree}
-:hidden:
-Datastore <self>
-```
-
-The state of a Kubernetes cluster is persisted in the [Datastore][].
-
-These topics cover how to configure the different datastore types that are
-supported by Canonical Kubernetes.
-
-```{toctree}
-:titlesonly:
-
-k8s-dqlite
-etcd
-external
-```
-
-<!-- LINKS -->
-
-[Datastore]: /snap/explanation/datastore/index
