+++ src/k8s/pkg/k8sd/app/hooks_join.go
@@ -118,9 +118,6 @@ func (a *App) onPostJoin(s *state.State, initConfig map[string]string) (rerr err
 		return fmt.Errorf("failed to get IP address(es) from ServiceCIDR %q: %w", cfg.Network.GetServiceCIDR(), err)
 	}
 
-	// Certificates
-	extraIPs, extraNames := utils.SplitIPAndDNSSANs(joinConfig.ExtraSANS)
-
 	switch cfg.Datastore.GetType() {
 	case "k8s-dqlite":
 		certificates := pki.NewK8sDqlitePKI(pki.K8sDqlitePKIOpts{
@@ -136,30 +133,6 @@ func (a *App) onPostJoin(s *state.State, initConfig map[string]string) (rerr err
 		if _, err := setup.EnsureK8sDqlitePKI(snap, certificates); err != nil {
 			return fmt.Errorf("failed to write k8s-dqlite certificates: %w", err)
 		}
-	case "etcd":
-		certificates := pki.NewEtcdPKI(pki.EtcdPKIOpts{
-			Hostname: s.Name(),
-			IPSANs:   append([]net.IP{nodeIP}, extraIPs...),
-			DNSSANs:  append([]string{s.Name()}, extraNames...),
-			Years:    20,
-		})
-
-		certificates.CACert = cfg.Datastore.GetEtcdCACert()
-		certificates.CAKey = cfg.Datastore.GetEtcdCAKey()
-		certificates.ServerCert = joinConfig.GetEtcdServerCert()
-		certificates.ServerKey = joinConfig.GetEtcdServerKey()
-		certificates.ServerPeerCert = joinConfig.GetEtcdServerPeerCert()
-		certificates.ServerPeerKey = joinConfig.GetEtcdServerPeerKey()
-		certificates.APIServerClientCert = cfg.Datastore.GetEtcdAPIServerClientCert()
-		certificates.APIServerClientKey = cfg.Datastore.GetEtcdAPIServerClientKey()
-
-		if err := certificates.CompleteCertificates(); err != nil {
-			return fmt.Errorf("failed to initialize etcd certificates: %w", err)
-		}
-		if _, err := setup.EnsureEtcdPKI(snap, certificates); err != nil {
-			return fmt.Errorf("failed to write etcd certificates: %w", err)
-		}
-
 	case "external":
 		certificates := &pki.ExternalDatastorePKI{
 			DatastoreCACert:     cfg.Datastore.GetExternalCACert(),
@@ -176,6 +149,8 @@ func (a *App) onPostJoin(s *state.State, initConfig map[string]string) (rerr err
 		return fmt.Errorf("unsupported datastore %s, must be one of %v", cfg.Datastore.GetType(), setup.SupportedDatastores)
 	}
 
+	// Certificates
+	extraIPs, extraNames := utils.SplitIPAndDNSSANs(joinConfig.ExtraSANS)
 	certificates := pki.NewControlPlanePKI(pki.ControlPlanePKIOpts{
 		Hostname:                  s.Name(),
 		IPSANs:                    append(append([]net.IP{nodeIP}, serviceIPs...), extraIPs...),
@@ -233,13 +208,11 @@ func (a *App) onPostJoin(s *state.State, initConfig map[string]string) (rerr err
 
 	// Configure datastore
 	switch cfg.Datastore.GetType() {
-	case "external":
-		// no-op
 	case "k8s-dqlite":
 		// TODO(neoaggelos): use cluster.GetInternalClusterMembers() instead
 		leader, err := s.Leader()
 		if err != nil {
-			return fmt.Errorf("failed to get microcluster leader: %w", err)
+			return fmt.Errorf("failed to get dqlite leader: %w", err)
 		}
 		members, err := leader.GetClusterMembers(ctx)
 		if err != nil {
@@ -247,36 +220,14 @@ func (a *App) onPostJoin(s *state.State, initConfig map[string]string) (rerr err
 		}
 		cluster := make([]string, len(members))
 		for _, member := range members {
-			cluster = append(cluster, utils.JoinHostPort(member.Address.Addr().String(), cfg.Datastore.GetK8sDqlitePort()))
+			cluster = append(cluster, fmt.Sprintf("%s:%d", member.Address.Addr(), cfg.Datastore.GetK8sDqlitePort()))
 		}
 
-		address := utils.JoinHostPort(nodeIP.String(), cfg.Datastore.GetK8sDqlitePort())
+		address := fmt.Sprintf("%s:%d", nodeIP.String(), cfg.Datastore.GetK8sDqlitePort())
 		if err := setup.K8sDqlite(snap, address, cluster, joinConfig.ExtraNodeK8sDqliteArgs); err != nil {
 			return fmt.Errorf("failed to configure k8s-dqlite with address=%s cluster=%v: %w", address, cluster, err)
 		}
-	case "etcd":
-		leader, err := s.Leader()
-		if err != nil {
-			return fmt.Errorf("failed to get microcluster leader: %w", err)
-		}
-		members, err := leader.GetClusterMembers(s.Context)
-		if err != nil {
-			return fmt.Errorf("failed to get microcluster members: %w", err)
-		}
-		clientURLs := make([]string, 0, len(members)-1)
-		for _, member := range members {
-			if member.Name == s.Name() {
-				// skip self
-				continue
-			}
-			clientURLs = append(clientURLs, fmt.Sprintf("https://%s", utils.JoinHostPort(member.Address.Addr().String(), cfg.Datastore.GetEtcdPort())))
-		}
-
-		clientURL := fmt.Sprintf("https://%s", utils.JoinHostPort(nodeIP.String(), cfg.Datastore.GetEtcdPort()))
-		peerURL := fmt.Sprintf("https://%s", utils.JoinHostPort(nodeIP.String(), cfg.Datastore.GetEtcdPeerPort()))
-		if err := setup.Etcd(snap, s.Name(), clientURL, peerURL, clientURLs, joinConfig.ExtraNodeK8sDqliteArgs); err != nil {
-			return fmt.Errorf("failed to configure etcd with peerURL=%s cluster=%v: %w", peerURL, clientURLs, err)
-		}
+	case "external":
 	default:
 		return fmt.Errorf("unsupported datastore %s, must be one of %v", cfg.Datastore.GetType(), setup.SupportedDatastores)
 	}
@@ -307,7 +258,7 @@ func (a *App) onPostJoin(s *state.State, initConfig map[string]string) (rerr err
 	if err := setup.KubeScheduler(snap, joinConfig.ExtraNodeKubeSchedulerArgs); err != nil {
 		return fmt.Errorf("failed to configure kube-scheduler: %w", err)
 	}
-	if err := setup.KubeAPIServer(snap, cfg.Network.GetServiceCIDR(), s.Address().Path("1.0", "kubernetes", "auth", "webhook").String(), true, cfg.Datastore, cfg.APIServer.GetAuthorizationMode(), s.Address().Hostname(), joinConfig.ExtraNodeKubeAPIServerArgs); err != nil {
+	if err := setup.KubeAPIServer(snap, cfg.Network.GetServiceCIDR(), s.Address().Path("1.0", "kubernetes", "auth", "webhook").String(), true, cfg.Datastore, cfg.APIServer.GetAuthorizationMode(), joinConfig.ExtraNodeKubeAPIServerArgs); err != nil {
 		return fmt.Errorf("failed to configure kube-apiserver: %w", err)
 	}
 
@@ -353,15 +304,6 @@ func (a *App) onPreRemove(s *state.State, force bool) (rerr error) {
 		rerr = nil
 	}()
 
-	// NOTE(neoaggelos): this is not sufficient, as it leaves behind other cluster resources tied to the node
-	c, err := snap.KubernetesClient("")
-	if err != nil {
-		return fmt.Errorf("failed to create Kubernetes client: %w", err)
-	}
-	if err := c.DeleteNode(s.Context, s.Name()); err != nil {
-		return fmt.Errorf("failed to remove k8s node %q: %w", s.Name(), err)
-	}
-
 	cfg, err := databaseutil.GetClusterConfig(s.Context, s)
 	if err != nil {
 		return fmt.Errorf("failed to retrieve k8sd cluster config: %w", err)
@@ -369,25 +311,27 @@ func (a *App) onPreRemove(s *state.State, force bool) (rerr error) {
 
 	// configure datastore
 	switch cfg.Datastore.GetType() {
-	case "external":
-		// no-op
 	case "k8s-dqlite":
 		client, err := snap.K8sDqliteClient(s.Context)
 		if err != nil {
 			return fmt.Errorf("failed to create k8s-dqlite client: %w", err)
 		}
 
-		nodeAddress := utils.JoinHostPort(s.Address().Hostname(), cfg.Datastore.GetK8sDqlitePort())
+		nodeAddress := net.JoinHostPort(s.Address().Hostname(), fmt.Sprintf("%d", cfg.Datastore.GetK8sDqlitePort()))
 		if err := client.RemoveNodeByAddress(s.Context, nodeAddress); err != nil {
 			return fmt.Errorf("failed to remove node with address %s from k8s-dqlite cluster: %w", nodeAddress, err)
 		}
-	case "etcd":
-		client := snap.EtcdClient()
-		nodeAddress := fmt.Sprintf("https://%s", utils.JoinHostPort(s.Address().Hostname(), cfg.Datastore.GetEtcdPeerPort()))
-		if err := client.RemoveNodeByAddress(s.Context, nodeAddress); err != nil {
-			return fmt.Errorf("failed to remove node with address %s from etcd cluster: %w", nodeAddress, err)
-		}
+	case "external":
 	default:
+	}
+
+	c, err := snap.KubernetesClient("")
+	if err != nil {
+		return fmt.Errorf("failed to create Kubernetes client: %w", err)
+	}
+
+	if err := c.DeleteNode(s.Context, s.Name()); err != nil {
+		return fmt.Errorf("failed to remove k8s node %q: %w", s.Name(), err)
 	}
 
 	return nil
