+++ src/k8s/pkg/k8sd/app/hooks_start.go
@@ -36,33 +36,9 @@ func (a *App) onStart(s *state.State) error {
 
 	// start control plane config controller
 	if a.controlPlaneConfigController != nil {
-		go a.controlPlaneConfigController.Run(
-			s.Context,
-			func(ctx context.Context) (types.ClusterConfig, []string, error) {
-				cfg, err := databaseutil.GetClusterConfig(ctx, s)
-				if err != nil {
-					return types.ClusterConfig{}, nil, fmt.Errorf("failed to retrieve cluster config: %w", err)
-				}
-
-				if cfg.Datastore.GetType() != "etcd" {
-					return cfg, nil, nil
-				}
-				leader, err := s.Leader()
-				if err != nil {
-					return types.ClusterConfig{}, nil, fmt.Errorf("failed to retrieve cluster leader: %w", err)
-				}
-				members, err := leader.GetClusterMembers(ctx)
-				if err != nil {
-					return types.ClusterConfig{}, nil, fmt.Errorf("failed to retrieve cluster members: %w", err)
-				}
-				nodeIPs := make([]string, 0, len(members))
-				for _, member := range members {
-					nodeIPs = append(nodeIPs, member.Address.Addr().String())
-				}
-
-				return cfg, nodeIPs, nil
-			},
-		)
+		go a.controlPlaneConfigController.Run(s.Context, func(ctx context.Context) (types.ClusterConfig, error) {
+			return databaseutil.GetClusterConfig(ctx, s)
+		})
 	}
 
 	// start update node config controller
