+++ src/k8s/pkg/k8sd/controllers/control_plane_configuration_test.go
@@ -2,12 +2,11 @@ package controllers_test
 
 import (
 	"context"
+	"github.com/canonical/k8s/pkg/utils"
 	"os"
 	"path"
 	"testing"
 	"time"
-
-	"github.com/canonical/k8s/pkg/utils"
 
 	"github.com/canonical/k8s/pkg/k8sd/controllers"
 	"github.com/canonical/k8s/pkg/k8sd/setup"
@@ -21,12 +20,7 @@ import (
 const channelSendTimeout = 100 * time.Millisecond
 
 type configProvider struct {
-	config  types.ClusterConfig
-	nodeIPs []string
-}
-
-func (c *configProvider) getConfigAndNodeIPs(ctx context.Context) (types.ClusterConfig, []string, error) {
-	return c.config, c.nodeIPs, nil
+	config types.ClusterConfig
 }
 
 func (c *configProvider) getConfig(ctx context.Context) (types.ClusterConfig, error) {
@@ -40,7 +34,6 @@ func TestControlPlaneConfigController(t *testing.T) {
 		s := &mock.Snap{
 			Mock: mock.Mock{
 				EtcdPKIDir:          path.Join(dir, "etcd-pki"),
-				KubernetesPKIDir:    path.Join(dir, "kube-pki"),
 				ServiceArgumentsDir: path.Join(dir, "args"),
 				UID:                 os.Getuid(),
 				GID:                 os.Getgid(),
@@ -54,12 +47,10 @@ func TestControlPlaneConfigController(t *testing.T) {
 		defer cancel()
 
 		triggerCh := make(chan time.Time)
-		configProvider := &configProvider{
-			nodeIPs: []string{"10.0.0.1", "10.0.0.2"},
-		}
+		configProvider := &configProvider{}
 
 		ctrl := controllers.NewControlPlaneConfigurationController(s, func() {}, triggerCh)
-		go ctrl.Run(ctx, configProvider.getConfigAndNodeIPs)
+		go ctrl.Run(ctx, configProvider.getConfig)
 
 		for _, tc := range []struct {
 			name   string
@@ -181,21 +172,6 @@ func TestControlPlaneConfigController(t *testing.T) {
 				},
 				expectServiceRestarts: []string{"kube-apiserver", "kube-controller-manager"},
 			},
-			{
-				name: "Etcd",
-				config: types.ClusterConfig{
-					Datastore: types.Datastore{
-						Type:     utils.Pointer("etcd"),
-						EtcdPort: utils.Pointer(12379),
-					},
-				},
-				expectKubeAPIServerArgs: map[string]string{
-					"--etcd-servers":  "https://10.0.0.1:12379,https://10.0.0.2:12379",
-					"--etcd-cafile":   path.Join(dir, "etcd-pki", "ca.crt"),
-					"--etcd-certfile": path.Join(dir, "kube-pki", "apiserver-etcd-client.crt"),
-					"--etcd-keyfile":  path.Join(dir, "kube-pki", "apiserver-etcd-client.key"),
-				},
-			},
 		} {
 			t.Run(tc.name, func(t *testing.T) {
 				g := NewWithT(t)
@@ -280,7 +256,7 @@ func TestControlPlaneConfigController(t *testing.T) {
 		configProvider := &configProvider{}
 
 		ctrl := controllers.NewControlPlaneConfigurationController(s, func() {}, triggerCh)
-		go ctrl.Run(ctx, configProvider.getConfigAndNodeIPs)
+		go ctrl.Run(ctx, configProvider.getConfig)
 
 		// mark as worker node
 		g.Expect(snaputil.MarkAsWorkerNode(s, true)).To(Succeed())
