+++ src/k8s/pkg/k8sd/pki/k8sdqlite.go
@@ -24,6 +24,7 @@ type K8sDqlitePKIOpts struct {
 	IPSANs            []net.IP
 	Years             int
 	AllowSelfSignedCA bool
+	Datastore         string
 }
 
 func NewK8sDqlitePKI(opts K8sDqlitePKIOpts) *K8sDqlitePKI {
