+++ src/k8s/pkg/k8sd/setup/directories.go
@@ -16,7 +16,6 @@ func EnsureAllDirectories(snap snap.Snap) error {
 		snap.ContainerdExtraConfigDir(),
 		snap.ContainerdRegistryConfigDir(),
 		snap.K8sDqliteStateDir(),
-		snap.EtcdDir(),
 		snap.KubernetesConfigDir(),
 		snap.KubernetesPKIDir(),
 		snap.EtcdPKIDir(),
