+++ src/k8s/pkg/k8sd/types/cluster_config_datastore.go
@@ -3,10 +3,7 @@ package types
 import (
 	"fmt"
 	"path"
-	"slices"
 	"strings"
-
-	"github.com/canonical/k8s/pkg/utils"
 )
 
 type Datastore struct {
@@ -20,13 +17,6 @@ type Datastore struct {
 	ExternalCACert     *string   `json:"external-ca-crt,omitempty"`
 	ExternalClientCert *string   `json:"external-client-crt,omitempty"`
 	ExternalClientKey  *string   `json:"external-client-key,omitempty"`
-
-	EtcdCACert              *string `json:"etcd-ca-crt,omitempty"`
-	EtcdCAKey               *string `json:"etcd-ca-key,omitempty"`
-	EtcdAPIServerClientCert *string `json:"etcd-apiserver-client-crt,omitempty"`
-	EtcdAPIServerClientKey  *string `json:"etcd-apiserver-client-key,omitempty"`
-	EtcdPort                *int    `json:"etcd-port,omitempty"`
-	EtcdPeerPort            *int    `json:"etcd-peer-port,omitempty"`
 }
 
 func (c Datastore) GetType() string               { return getField(c.Type) }
@@ -37,28 +27,17 @@ func (c Datastore) GetExternalServers() []string  { return getField(c.ExternalSe
 func (c Datastore) GetExternalCACert() string     { return getField(c.ExternalCACert) }
 func (c Datastore) GetExternalClientCert() string { return getField(c.ExternalClientCert) }
 func (c Datastore) GetExternalClientKey() string  { return getField(c.ExternalClientKey) }
-func (c Datastore) GetEtcdCACert() string         { return getField(c.EtcdCACert) }
-func (c Datastore) GetEtcdCAKey() string          { return getField(c.EtcdCAKey) }
-func (c Datastore) GetEtcdAPIServerClientCert() string {
-	return getField(c.EtcdAPIServerClientCert)
-}
-func (c Datastore) GetEtcdAPIServerClientKey() string {
-	return getField(c.EtcdAPIServerClientKey)
-}
-func (c Datastore) GetEtcdPort() int     { return getField(c.EtcdPort) }
-func (c Datastore) GetEtcdPeerPort() int { return getField(c.EtcdPeerPort) }
-func (c Datastore) Empty() bool          { return c == Datastore{} }
+func (c Datastore) Empty() bool                   { return c == Datastore{} }
 
 // DatastorePathsProvider is to avoid circular dependency for snap.Snap in Datastore.ToKubeAPIServerArguments()
 type DatastorePathsProvider interface {
-	KubernetesPKIDir() string
 	K8sDqliteStateDir() string
 	EtcdPKIDir() string
 }
 
 // ToKubeAPIServerArguments returns updateArgs, deleteArgs that can be used with snaputil.UpdateServiceArguments() for the kube-apiserver
 // according the datastore configuration.
-func (c Datastore) ToKubeAPIServerArguments(p DatastorePathsProvider, nodeIPs []string) (map[string]string, []string) {
+func (c Datastore) ToKubeAPIServerArguments(p DatastorePathsProvider) (map[string]string, []string) {
 	var (
 		updateArgs = make(map[string]string)
 		deleteArgs []string
@@ -87,22 +66,6 @@ func (c Datastore) ToKubeAPIServerArguments(p DatastorePathsProvider, nodeIPs []
 				deleteArgs = append(deleteArgs, loop.arg)
 			}
 		}
-	case "etcd":
-		updateArgs["--etcd-cafile"] = path.Join(p.EtcdPKIDir(), "ca.crt")
-		updateArgs["--etcd-certfile"] = path.Join(p.KubernetesPKIDir(), "apiserver-etcd-client.crt")
-		updateArgs["--etcd-keyfile"] = path.Join(p.KubernetesPKIDir(), "apiserver-etcd-client.key")
-
-		// Silently ignore an empty list of clientURLs and do not update the --etcd-servers argument.
-		if len(nodeIPs) == 0 {
-			break
-		}
-		clientURLs := make([]string, 0, len(nodeIPs))
-		for _, ip := range nodeIPs {
-			clientURLs = append(clientURLs, fmt.Sprintf("https://%s", utils.JoinHostPort(ip, c.GetEtcdPort())))
-		}
-		slices.Sort(clientURLs)
-
-		updateArgs["--etcd-servers"] = strings.Join(clientURLs, ",")
 	}
 
 	return updateArgs, deleteArgs
