+++ src/k8s/pkg/k8sd/types/cluster_config_datastore_convert_test.go
@@ -1,11 +1,11 @@
 package types_test
 
 import (
+	"github.com/canonical/k8s/pkg/utils"
 	"testing"
 
 	apiv1 "github.com/canonical/k8s/api/v1"
 	"github.com/canonical/k8s/pkg/k8sd/types"
-	"github.com/canonical/k8s/pkg/utils"
 	. "github.com/onsi/gomega"
 )
