+++ src/k8s/pkg/k8sd/types/cluster_config_datastore_test.go
@@ -1,9 +1,8 @@
 package types_test
 
 import (
-	"testing"
-
 	"github.com/canonical/k8s/pkg/utils"
+	"testing"
 
 	"github.com/canonical/k8s/pkg/k8sd/types"
 	"github.com/canonical/k8s/pkg/snap/mock"
@@ -71,7 +70,7 @@ func TestDatastoreToKubeAPIServerArguments(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			g := NewWithT(t)
 
-			update, delete := tc.config.ToKubeAPIServerArguments(snap, nil)
+			update, delete := tc.config.ToKubeAPIServerArguments(snap)
 			g.Expect(update).To(Equal(tc.expectUpdateArgs))
 			g.Expect(delete).To(Equal(tc.expectDeleteArgs))
 		})
