+++ src/k8s/pkg/k8sd/types/cluster_config_defaults_test.go
@@ -1,9 +1,8 @@
 package types_test
 
 import (
-	"testing"
-
 	"github.com/canonical/k8s/pkg/utils"
+	"testing"
 
 	"github.com/canonical/k8s/pkg/k8sd/types"
 	. "github.com/onsi/gomega"
@@ -27,8 +26,6 @@ func TestSetDefaults(t *testing.T) {
 		Datastore: types.Datastore{
 			Type:          utils.Pointer("k8s-dqlite"),
 			K8sDqlitePort: utils.Pointer(9000),
-			EtcdPort:      utils.Pointer(2379),
-			EtcdPeerPort:  utils.Pointer(2380),
 		},
 		Kubelet: types.Kubelet{
 			ClusterDomain: utils.Pointer("cluster.local"),
