+++ src/k8s/pkg/k8sd/types/cluster_config_merge_test.go
@@ -95,16 +95,6 @@ func TestMergeClusterConfig(t *testing.T) {
 		generateMergeClusterConfigTestCases("Datastore/ExternalCACert", true, "v1", "v2", func(c *types.ClusterConfig, v any) { c.Datastore.ExternalCACert = utils.Pointer(v.(string)) }),
 		generateMergeClusterConfigTestCases("Datastore/ExternalClientCert", true, "v1", "v2", func(c *types.ClusterConfig, v any) { c.Datastore.ExternalClientCert = utils.Pointer(v.(string)) }),
 		generateMergeClusterConfigTestCases("Datastore/ExternalClientKey", true, "v1", "v2", func(c *types.ClusterConfig, v any) { c.Datastore.ExternalClientKey = utils.Pointer(v.(string)) }),
-		generateMergeClusterConfigTestCases("Datastore/EtcdCACert", false, "v1", "v2", func(c *types.ClusterConfig, v any) { c.Datastore.EtcdCACert = utils.Pointer(v.(string)) }),
-		generateMergeClusterConfigTestCases("Datastore/EtcdCAKey", false, "v1", "v2", func(c *types.ClusterConfig, v any) { c.Datastore.EtcdCAKey = utils.Pointer(v.(string)) }),
-		generateMergeClusterConfigTestCases("Datastore/EtcdAPIServerClientCert", true, "v1", "v2", func(c *types.ClusterConfig, v any) {
-			c.Datastore.EtcdAPIServerClientCert = utils.Pointer(v.(string))
-		}),
-		generateMergeClusterConfigTestCases("Datastore/EtcdAPIServerClientKey", true, "v1", "v2", func(c *types.ClusterConfig, v any) {
-			c.Datastore.EtcdAPIServerClientKey = utils.Pointer(v.(string))
-		}),
-		generateMergeClusterConfigTestCases("Datastore/EtcdPort", false, 2379, 12379, func(c *types.ClusterConfig, v any) { c.Datastore.EtcdPort = utils.Pointer(v.(int)) }),
-		generateMergeClusterConfigTestCases("Datastore/EtcdPeerPort", false, 2380, 12380, func(c *types.ClusterConfig, v any) { c.Datastore.EtcdPeerPort = utils.Pointer(v.(int)) }),
 		generateMergeClusterConfigTestCases("Network/Enable", true, true, false, func(c *types.ClusterConfig, v any) { c.Network.Enabled = utils.Pointer(v.(bool)) }),
 		generateMergeClusterConfigTestCases("Network/Disable", true, false, true, func(c *types.ClusterConfig, v any) { c.Network.Enabled = utils.Pointer(v.(bool)) }),
 		generateMergeClusterConfigTestCases("Network/PodCIDR", false, "10.1.0.0/16", "10.2.0.0/16", func(c *types.ClusterConfig, v any) { c.Network.PodCIDR = utils.Pointer(v.(string)) }),
