+++ src/k8s/pkg/snap/interface.go
@@ -4,7 +4,6 @@ import (
 	"context"
 
 	"github.com/canonical/k8s/pkg/client/dqlite"
-	"github.com/canonical/k8s/pkg/client/etcd"
 	"github.com/canonical/k8s/pkg/client/helm"
 	"github.com/canonical/k8s/pkg/client/k8sd"
 	"github.com/canonical/k8s/pkg/client/kubernetes"
@@ -46,7 +45,6 @@ type Snap interface {
 
 	K8sdStateDir() string      // /var/snap/k8s/common/var/lib/k8sd/state
 	K8sDqliteStateDir() string // /var/snap/k8s/common/var/lib/k8s-dqlite
-	EtcdDir() string           // /var/snap/k8s/common/var/lib/etcd
 
 	ServiceArgumentsDir() string   // /var/snap/k8s/common/args
 	ServiceExtraConfigDir() string // /var/snap/k8s/common/args/conf.d
@@ -59,8 +57,6 @@ type Snap interface {
 	HelmClient() helm.Client // admin helm client
 
 	K8sDqliteClient(ctx context.Context) (*dqlite.Client, error) // go-dqlite client for k8s-dqlite
-
-	EtcdClient() etcd.Client // client for the managed etcd cluster
 
 	K8sdClient(address string) (k8sd.Client, error) // k8sd client
