+++ src/k8s/pkg/utils/cidr.go
@@ -125,8 +125,3 @@ func ParseCIDRs(CIDRstring string) (string, string, error) {
 	}
 	return ipv4CIDR, ipv6CIDR, nil
 }
-
-// JoinHostPort is like net.JoinHostPort but accepts an int for the port number.
-func JoinHostPort(host string, port int) string {
-	return net.JoinHostPort(host, strconv.Itoa(port))
-}
