+++ tests/integration/tests/test_clustering.py
@@ -31,3 +31,36 @@ def test_control_plane_nodes(instances: List[harness.Instance]):
     assert (
         nodes[0]["metadata"]["name"] == cluster_node.id
     ), f"only {cluster_node.id} should be left in cluster"
+
+
+@pytest.mark.node_count(3)
+def test_worker_nodes(instances: List[harness.Instance]):
+    cluster_node = instances[0]
+    joining_node = instances[1]
+    other_joining_node = instances[2]
+
+    join_token = util.get_join_token(cluster_node, joining_node, "--worker")
+    join_token_2 = util.get_join_token(cluster_node, other_joining_node, "--worker")
+
+    assert join_token != join_token_2
+
+    util.join_cluster(joining_node, join_token)
+
+    util.join_cluster(other_joining_node, join_token_2)
+
+    util.wait_until_k8s_ready(cluster_node, instances)
+    nodes = util.ready_nodes(cluster_node)
+    assert len(nodes) == 3, "workers should have joined cluster"
+
+    assert "control-plane" in util.get_local_node_status(cluster_node)
+    assert "worker" in util.get_local_node_status(joining_node)
+    assert "worker" in util.get_local_node_status(other_joining_node)
+
+    cluster_node.exec(["k8s", "remove-node", joining_node.id])
+    nodes = util.ready_nodes(cluster_node)
+    assert len(nodes) == 2, "worker should have been removed from cluster"
+    assert cluster_node.id in [
+        node["metadata"]["name"] for node in nodes
+    ] and other_joining_node.id in [
+        node["metadata"]["name"] for node in nodes
+    ], f"only {cluster_node.id} should be left in cluster"
