+++ tests/integration/tests/test_etcd.py
@@ -1,84 +1,101 @@
 #
 # Copyright 2024 Canonical, Ltd.
 #
+import json
 import logging
 from typing import List
 
 import pytest
 import yaml
 from test_util import harness, util
+from test_util.etcd import EtcdCluster
 
 LOG = logging.getLogger(__name__)
 
 
-@pytest.mark.node_count(2)
+@pytest.mark.node_count(1)
+@pytest.mark.etcd_count(1)
 @pytest.mark.disable_k8s_bootstrapping()
-def test_etcd(instances: List[harness.Instance]):
-    cluster_node = instances[0]
-    joining_node = instances[1]
+def test_etcd(instances: List[harness.Instance], etcd_cluster: EtcdCluster):
+    k8s_instance = instances[0]
 
     bootstrap_conf = yaml.safe_dump(
         {
             "cluster-config": {"network": {"enabled": True}, "dns": {"enabled": True}},
-            "datastore-type": "etcd",
+            "datastore-type": "external",
+            "datastore-servers": etcd_cluster.client_urls,
+            "datastore-ca-crt": etcd_cluster.ca_cert,
+            "datastore-client-crt": etcd_cluster.cert,
+            "datastore-client-key": etcd_cluster.key,
         }
     )
 
-    cluster_node.exec(
+    k8s_instance.exec(
         ["dd", "of=/root/config.yaml"],
         input=str.encode(bootstrap_conf),
     )
 
-    cluster_node.exec(["k8s", "bootstrap", "--file", "/root/config.yaml"])
-    util.wait_for_dns(cluster_node)
-    util.wait_for_network(cluster_node)
+    k8s_instance.exec(["k8s", "bootstrap", "--file", "/root/config.yaml"])
+    util.wait_for_dns(k8s_instance)
+    util.wait_for_network(k8s_instance)
 
-    join_token = util.get_join_token(cluster_node, joining_node)
-    util.join_cluster(joining_node, join_token)
+    p = k8s_instance.exec(
+        ["systemctl", "is-active", "--quiet", "snap.k8s.k8s-dqlite"], check=False
+    )
+    assert p.returncode != 0, "k8s-dqlite service is still active"
 
-    util.wait_until_k8s_ready(cluster_node, instances)
-    nodes = util.ready_nodes(cluster_node)
-    assert len(nodes) == 2, "node should have joined cluster"
+    LOG.info("Add new etcd nodes")
+    etcd_cluster.add_nodes(2)
 
-    assert "control-plane" in util.get_local_node_status(cluster_node)
-    assert "control-plane" in util.get_local_node_status(joining_node)
+    # Update server-urls in cluster
+    body = {
+        "datastore": {
+            "type": "external",
+            "servers": etcd_cluster.client_urls,
+            "ca-crt": etcd_cluster.ca_cert,
+            "client-crt": etcd_cluster.cert,
+            "client-key": etcd_cluster.key,
+        }
+    }
+    k8s_instance.exec(
+        [
+            "curl",
+            "-XPUT",
+            "--header",
+            "Content-Type: application/json",
+            "--data",
+            json.dumps(body),
+            "--unix-socket",
+            "/var/snap/k8s/common/var/lib/k8sd/state/control.socket",
+            "http://localhost/1.0/k8sd/cluster/config",
+        ]
+    )
 
-    cluster_node.exec(["k8s", "remove-node", joining_node.id])
-    nodes = util.ready_nodes(cluster_node)
-    assert len(nodes) == 1, "node should have been removed from cluster"
-    assert (
-        nodes[0]["metadata"]["name"] == cluster_node.id
-    ), f"only {cluster_node.id} should be left in cluster"
+    # check that we can still connect to the kubernetes cluster
+    util.stubbornly(retries=10, delay_s=2).on(k8s_instance).exec(
+        ["k8s", "kubectl", "get", "pods", "-A"]
+    )
 
+    # Changing the datastore back to k8s-dqlite after using the external datastore should fail.
+    body = {
+        "datastore": {
+            "type": "k8s-dqlite",
+        }
+    }
 
-@pytest.mark.node_count(3)
-def test_worker_nodes(instances: List[harness.Instance]):
-    cluster_node = instances[0]
-    joining_node = instances[1]
-    other_joining_node = instances[2]
-
-    join_token = util.get_join_token(cluster_node, joining_node, "--worker")
-    join_token_2 = util.get_join_token(cluster_node, other_joining_node, "--worker")
-
-    assert join_token != join_token_2
-
-    util.join_cluster(joining_node, join_token)
-
-    util.join_cluster(other_joining_node, join_token_2)
-
-    util.wait_until_k8s_ready(cluster_node, instances)
-    nodes = util.ready_nodes(cluster_node)
-    assert len(nodes) == 3, "workers should have joined cluster"
-
-    assert "control-plane" in util.get_local_node_status(cluster_node)
-    assert "worker" in util.get_local_node_status(joining_node)
-    assert "worker" in util.get_local_node_status(other_joining_node)
-
-    cluster_node.exec(["k8s", "remove-node", joining_node.id])
-    nodes = util.ready_nodes(cluster_node)
-    assert len(nodes) == 2, "worker should have been removed from cluster"
-    assert cluster_node.id in [
-        node["metadata"]["name"] for node in nodes
-    ] and other_joining_node.id in [
-        node["metadata"]["name"] for node in nodes
-    ], f"only {cluster_node.id} should be left in cluster"
+    resp = k8s_instance.exec(
+        [
+            "curl",
+            "-XPUT",
+            "--header",
+            "Content-Type: application/json",
+            "--data",
+            json.dumps(body),
+            "--unix-socket",
+            "/var/snap/k8s/common/var/lib/k8sd/state/control.socket",
+            "http://localhost/1.0/k8sd/cluster/config",
+        ],
+        capture_output=True,
+    )
+    response = json.loads(resp.stdout.decode())
+    assert response["error_code"] == 400, "changing the datastore type should fail"
